fixed folder structure and precommit errors
