--- src/subjugator/command/.vscode/c_cpp_properties.json
@@ -1,26 +1,0 @@
-{
-  "configurations": [
-    {
-      "browse": {
-        "databaseFilename": "${default}",
-        "limitSymbolsToIncludedHeaders": false
-      },
-      "includePath": [
-        "/home/aarush-tripathi/mil2/install/subjugator_msgs/include/**",
-        "/home/aarush-tripathi/mil2/install/subjugator_gazebo/include/**",
-        "/home/aarush-tripathi/mil2/install/mil_passive_sonar/include/**",
-        "/opt/ros/jazzy/include/**",
-        "/home/aarush-tripathi/mil2/src/mil_common/mil_tools/include/**",
-        "/home/aarush-tripathi/mil2/src/subjugator/command/subjugator_joy/src/joy_subjugator/include/**",
-        "/home/aarush-tripathi/mil2/src/subjugator/simulation/subjugator_gazebo/include/**",
-        "/usr/include/**"
-      ],
-      "name": "ROS",
-      "intelliSenseMode": "gcc-x64",
-      "compilerPath": "/usr/bin/gcc",
-      "cStandard": "gnu11",
-      "cppStandard": "c++14"
-    }
-  ],
-  "version": 4
-}
