--- src/subjugator/command/subjugator_joy/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "python.autoComplete.extraPaths": [
-        "/opt/ros/jazzy/lib/python3.12/site-packages"
-    ],
-    "python.analysis.extraPaths": [
-        "/opt/ros/jazzy/lib/python3.12/site-packages"
-    ]
-}
